Allow setting different namespace root
Currently the root for any loaded namespace is the loader instance. But
someone may want to set autoloads on a nother (named or anonymous)
module, and there is no reason Im should not support this.

With this change, it is now possible to pass a "root" when initializing
a loader. This root (a module) is then used in place of the loader
itself as the root of the (auto)loaded namespace.